updated properties

+++ mo_class.h
@@ -9,12 +9,12 @@
 class MO {
 	private:
 		int nr;
-		int occ;
+		double occ;
 		double ener;
 		std::vector<double> coefficients;
 	public:
 		MO();
-		MO (int number,float occupation,double energy);
+		MO (int number,double occupation,double energy);
 		bool push_back_coef(double val, const int nex);
 		bool erase_coef(int nr, const int nex);
 		double get_coefficient(int nr, bool debug) const;
@@ -22,7 +22,7 @@ class MO {
 		bool change_coefficient(int nr, double value, bool debug);
 		void set_nr(int inr) { nr=inr; };
 		void set_occ(int iocc) { occ=iocc; };
-		int get_occ() { return occ; };
+		double get_occ() { return occ; };
 		void set_ener(double iener) {ener=iener;};
 		int get_primitive_count() const {return coefficients.size();};
 		std::string hdr();
